requirements.txt
oci==2.129.2
requests==2.32.3
python-dotenv==1.0.1